requirements.txt
streamlit
ultralytics
opencv-python
Pillow
numpy
python-dotenv